evaluation_model/anygpt/requirements_anygpt.txt
absl-py==2.1.0
accelerate==0.32.1
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
beartype==0.16.4
blinker==1.8.2
cachetools==5.3.3
cdifflib==1.2.6
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
clean-text==0.6.0
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
datasets==2.14.7
decorator==5.1.1
diffusers==0.20.2
dill==0.3.7
dnspython==2.6.1
editdistance==0.8.1
einops==0.8.0
email_validator==2.2.0
emoji==1.7.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.2.0
google-auth==2.32.0
google-auth-oauthlib==1.2.1
gradio==3.50.0
gradio_client==0.6.1
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
hydra-core==1.3.2
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.3.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.0
lion-pytorch==0.2.2
llvmlite==0.43.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.15
nemo_text_processing==1.0.2
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==4.23.4
psutil==6.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pynini==2.1.5
pyparsing==3.1.2
pyrootutils==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
sacremoses==0.1.1
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.11.4
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
speechtokenizer==0.1.2
starlette==0.37.2
sympy==1.13.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
timm==1.0.7
tokenizers==0.15.2
toolz==0.12.1
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.37.0
triton==2.1.0
typeguard==4.3.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.3
wget==3.2
wrapt==1.16.0
xformers==0.0.22.post7
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
